pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kais-sdk"
version = "0.1.0"
description = "kAIs SDK — connect your apps to kAIs cells via NATS"
requires-python = ">=3.10"
dependencies = ["nats-py>=2.7.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.hatch.build.targets.wheel]
packages = ["src/kais"]
